ci: add workflow pipelines for automated testing and debug package generation

+++ .github/make_debug_zip.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 
-mkdir output
-cp app/build/outputs/apk/debug/PixelXpert-signed.apk MagiskModBase/system/priv-app/PixelXpert/PixelXpert.apk
+mkdir -p output
+if [ -f app/build/outputs/apk/debug/PixelXpert-signed.apk ]; then
+  cp app/build/outputs/apk/debug/PixelXpert-signed.apk MagiskModBase/system/priv-app/PixelXpert/PixelXpert.apk
+else
+  cp app/build/outputs/apk/debug/PixelXpert.apk MagiskModBase/system/priv-app/PixelXpert/PixelXpert.apk
+fi
 cd MagiskModBase
 
 FILENAME="PixelXpert.zip"
